requirements.txt
discord.py
python-dotenv
yt-dlp
ffmpeg
requests
